Substantial refactoring and docs
Most noticeable change is simplifying the example that gets initialized in a new R project to use sb_create_app instead of including the shinydashboard code.

One breaking change to my knowledge: this changes the API of sb_create_app to take a _site.yml so that it can't take a configuration (always reads it from _site.yml)

Adds functions sb_ui and sb_server, breaking out the code in sb_create_app a bit so that it's easier to customize and build on top of.

Draft at supporting pinned data in YAML file with the !pin tag. Also supports !global tag to refer to a variable in data_global.

Added detail to a lot of the docs. Roxygen2 update.